refactor: controller, inputer, viewer 통합 작업2
