--- src/main/java/lottogame/controller/purchase/spi/PurchaseInputer.java
@@ -1,7 +1,0 @@
-package lottogame.controller.purchase.spi;
-
-public interface PurchaseInputer {
-
-    int inputMoney();
-
-}
