--- src/main/java/lottogame/controller/purchase/spi/PurchaseViewer.java
@@ -1,10 +1,0 @@
-package lottogame.controller.purchase.spi;
-
-import java.util.List;
-import lottogame.domain.LottoTicket;
-
-public interface PurchaseViewer {
-
-    void draw(List<LottoTicket> lottoTickets);
-
-}
