--- src/main/java/lottogame/controller/resultcheck/spi/ResultCheckInputer.java
@@ -1,11 +1,0 @@
-package lottogame.controller.resultcheck.spi;
-
-import java.util.Set;
-
-public interface ResultCheckInputer {
-
-    Set<Integer> inputWinningLottoNumbers();
-
-    Integer inputBonusLottoNumber();
-    
-}
