--- src/main/java/lottogame/controller/resultcheck/spi/ResultCheckViewer.java
@@ -1,9 +1,0 @@
-package lottogame.controller.resultcheck.spi;
-
-import lottogame.service.response.LottoCheckResponse;
-
-public interface ResultCheckViewer {
-
-    void draw(LottoCheckResponse lottoCheckResponse);
-
-}
